backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
pandas==2.2.3
openpyxl==3.1.5
google-genai==0.3.0
python-multipart==0.0.12
jinja2==3.1.4
xlsxwriter==3.2.0
